requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.3.7
anthropic==0.7.7
python-dotenv==1.0.0
aiofiles==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
torch==2.1.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
aiohttp==3.9.1
pydantic-settings==2.0.0
huggingface-hub==0.24.0
transformers==4.37.2
tiktoken==0.3.3